requirements.txt
python-dotenv
httpx[http2]
beautifulsoup4
olx-db-wonsky
pydantic-settings
pytz

langchain-groq
psycopg2-binary
